backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
pillow==12.2.0
psycopg2-binary==2.9.12
sqlparse==0.5.5
tzdata==2026.2
